Removal of a backdoor allowing to call functions, import modules and read and write attrs.
This lib claims:

> Simple **Safe** Sandboxed Extensible Expression Evaluator

So it must be secure by default. But it is not possible to make it secure in presence of such dangerous functions. So this commit gets rid of them.

https://nedbatchelder.com/blog/201206/eval_really_is_dangerous.html claims

>The problem with all of these attempts to protect eval() is that they are blacklists. They explicitly remove things that could be dangerous. That is a losing battle because if there’s just one item left off the list, you can attack the system.

and I completely aggree with it.

If a user needs to allow these AST nodes in an insecure way (which defeats purpose of this lib), he can do it explicitly. The functions are still in place, all he needs is to subclass the class and add them to the dict.

If a user needs to allow them in a more secure way, he will implement them himself knowing the specifics of the code he calls and allowing only the API he is sure must be called by the code and in a more secure way.